digitalocean/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
transformers>=4.36.0
torch>=2.4.0
accelerate>=0.24.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
Pillow>=10.0.0
torchvision>=0.19.0
